fix(kernel): scope per-agent cap docs to trigger dispatch, tighten resolver
Addresses review feedback on #3210:

- H1: scope docstrings (manifest field, kernel field, arch doc, CLAUDE.md)
  to the trigger dispatcher only — agent_send / channel / cron paths
  do not acquire the per-agent semaphore.
- M1: tighten lifecycle wording — picking up a new cap requires kill
  and respawn (or daemon restart); in-place activate/status flip is a
  no-op for cap changes.
- M2: single registry read in agent_concurrency_for so cap and
  session_mode come from the same manifest snapshot (closes TOCTOU
  across a manifest swap).
- M3: type-align default_per_agent to usize, drop the as-cast in the
  resolver; schema golden regenerated (format: "uint").
- L3: add test_semaphore_for_lane_routes_each_variant with distinct
  capacities to catch a copy-paste bug in the lane match arm.
- L4: rework manifest doc — auto-clamps to 1 unless session_mode = "new";
  Some(0) silently floored to 1.
- L5: arch doc table gains a "Per-agent cap applies?" column; only
  trigger dispatch is yes.
- L2: tuning note added — both permits held across full agent loop.
- Nit: rename local trigger_lane → trigger_sem to stop shadowing the
  config field name.

+++ crates/librefang-types/src/config/types.rs
@@ -1866,8 +1866,9 @@ pub struct QueueConcurrencyConfig {
     /// not set `max_concurrent_invocations`. `1` reproduces the
     /// legacy per-agent-mutex serialization that pre-existed this
     /// knob — change deliberately. `0` is rewritten to `1` by
-    /// validation.
-    pub default_per_agent: u32,
+    /// validation. Typed `usize` to match the sibling lane fields and
+    /// to feed `Semaphore::new` without a cast.
+    pub default_per_agent: usize,
 }
 
 impl Default for QueueConcurrencyConfig {
